Index files edited and more files whitelisted in ajax.js and files.js. Navigation files edited for validation of navigation settings, Signed-off-by: Piyush Vijay <piyushvijay.1997@gmail.com>
