+++ .gitignore
@@ -55,3 +55,6 @@ composer.lock
 /vendor/
 # NPM
 /node_modules/
+# Javascript Bundle
+js/dist/
+js/lib/
