+++ js/src/consts/files.js
@@ -1,10 +1,18 @@
 const files = {
+    global: ['error_report', 'config', 'navigation'],
     server_privileges: ['server_privileges'],
     server_databases: ['server_databases'],
-    global: ['error_report', 'navigation'],
     server_status_advisor: ['server_status_advisor'],
     server_status_processes: ['server_status_processes'],
-    server_status_variables: ['server_status_variables']
+    server_status_variables: ['server_status_variables'],
+    server_plugins: ['server_plugins'],
+    server_status_queries: ['server_status_sorter', 'server_status_queries'],
+    server_status_monitor: ['server_status_monitor', 'server_status_sorter'],
+    server_variables: ['server_variables'],
+    server_user_groups: ['server_user_groups'],
+    server_replication: ['server_privileges', 'replication'],
+    server_export: ['export'],
+    server_import: ['import']
 };
 
 export default files;
